Retrieve GrieneisenTS data

Obtain longitudinal gut microbiome data in wild baboons from Grieneisen et al. (2021).

GrieneisenTSData()

baboongut()

Value

A TreeSummarizedExperiment object.

Details

The GrieneisenTS dataset contains 16,234 16S rRNA gene sequencing-based microbiome profiles from 585 baboon samples collected over 14 years to determine the heritability of the gut microbiome on various environmental factors such as diet, age, season. Each baboon had an average of 28 samples collected over 4.5 years. The data set can be used to investigate significance of longitudinal sampling at large sample sizes.

This data set contains the 613 most prevalent taxa with a phylogenetic tree.

Column metadata includes the following fields:

sample: Sample ID (character)

baboon_id: Baboon ID (factor)

collection_date: Sample collection date (date ; YYYY-MM-DD)

sex: Sex (factor; F/M)

age: Age (numeric)

social_group: Social group ID (factor)

group_size: Social group size (integer)

rain_month_mm: Rain per month(mm) (numeric)

season: Season (factor; dry/wet)

hydro_year: Hydro year (integer)

month: Month (integer)

readcount: Read count (numeric)

plate: Plate (factor)

post_pcr_dna_ng: Post PCR DNA(ng) (numeric)

diet_PC1: Diet Principal coordinate 1 (numeric)

diet_PC2: Diet Principal coordinate 2 (numeric)

diet_PC3: Diet Principal coordinate 3 (numeric)

diet_PC4: Diet Principal coordinate 4 (numeric)

diet_PC5: Diet Principal coordinate 5 (numeric)

diet_PC6: Diet Principal coordinate 6 (numeric)

diet_PC7: Diet Principal coordinate 7 (numeric)

diet_PC8: Diet Principal coordinate 8 (numeric)

diet_PC9: Diet Principal coordinate 9 (numeric)

diet_PC10: Diet Principal coordinate 10 (numeric)

diet_PC11: Diet Principal coordinate 11 (numeric)

diet_PC12: Diet Principal coordinate 12 (numeric)

diet_PC13: Diet Principal coordinate 13 (numeric)

diet_shannon_h: Dietary Shannon's H index (numeric)

asv_richness: Amplicon sequence variant (ASV) richness (integer)

asv_shannon_h: ASV Shannon's H index (numeric)

pc1_bc: Principal coordinate 1 Bray-Curtis dissimilarity (numeric)

pc2_bc: Principal coordinate 2 Bray-Curtis dissimilarity (numeric)

pc3_bc: Principal coordinate 3 Bray-Curtis dissimilarity (numeric)

pc4_bc: Principal coordinate 4 Bray-Curtis dissimilarity (numeric)

pc5_bc: Principal coordinate 5 Bray-Curtis dissimilarity (numeric)

Row metadata of the microbiome data contains taxonomic information on the Domain, Phylum, Class, Order, Family, Genus, and ASV levels.

The row tree consists of a phylogenetic tree build using sequence information of 613 taxa.

As reference sequences the ASV are provided.

References

Grieneisen et al. (2021): Gut microbiome heritability is nearly universal but environmentally contingent Science 373:6551 https://science.sciencemag.org/content/373/6551/181.full

Author

Yagmur Simsek and Leo Lahti

Examples

tse <- GrieneisenTSData()
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache