Retrieve HintikkaXO data

Obtain microbiota, metabolite and biomarker data from Hintikka et al. (2021).

HintikkaXOData()

Value

for HintikkaXOData a MultiAssayExperiment

is provided.

Details

The HintikkaXO dataset contains high-throughput profiling data from 40 rat samples, including 39 biomarkers, 38 metabolites (NMR), and 12706 OTUs from 318 species, measured from Cecum. This is diet comparison study with High/Low fat diet and xylo-oligosaccaride supplementation.

Column metadata is common for all experiments (microbiota, metabolites, biomarkers) and includes the following fields:

Sample: Sample ID (character)

Rat: Rat ID (factor)

Site: Site of measurement ("Cecum"); single value

Diet: Diet group (factor; combination of the Fat and XOS fields)

Fat: Fat in Diet (factor; Low/High)

XOS: XOS Diet Supplement (numeric; 0/1)

Row metadata of the microbiota data contains taxonomic information on the Phylum, Class, Order, Family, Genus, Species, and OTU levels.

Biomarker data contains 39 biomarkers.

Metabolite data contains 38 NMR metabolites.

All data are downloaded from ExperimentHub and cached for local re-use.

References

Hintikka L et al. (2021): Xylo-oligosaccharides in prevention of hepatic steatosis and adipose tissue inflammation: associating taxonomic and metabolomic patterns in fecal microbiotas with biclustering. International Journal of Environmental Research and Public Health 18(8):4049 https://doi.org/10.3390/ijerph18084049

Author

Leo Lahti

Examples

# Retrieve the MAE data
mae <- HintikkaXOData()
#> Warning: HintikkaXOData has been moved to the mia package.
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache
#> see ?microbiomeDataSets and browseVignettes('microbiomeDataSets') for documentation
#> loading from cache

# List the experiments in this MultiAssayExperiment object
print(experiments(mae))
#> ExperimentList class object of length 3:
#>  [1] microbiota: SummarizedExperiment with 12706 rows and 40 columns
#>  [2] metabolites: SummarizedExperiment with 38 rows and 40 columns
#>  [3] biomarkers: SummarizedExperiment with 39 rows and 40 columns

# colData for this MAE data object (print first few rows)
print(head(colData(mae)))
#> DataFrame with 6 rows and 6 columns
#>         Sample      Rat        Site     Diet      Fat       XOS
#>    <character> <factor> <character> <factor> <factor> <numeric>
#> C1          C1        1       Cecum High-fat     High         0
#> C2          C2        2       Cecum High-fat     High         0
#> C3          C3        3       Cecum High-fat     High         0
#> C4          C4        4       Cecum High-fat     High         0
#> C5          C5        5       Cecum High-fat     High         0
#> C6          C6        6       Cecum High-fat     High         0

# metabolite assay data
nmr <- assays(mae[["metabolites"]])$nmr

# biomarker assay data
bm <- assays(mae[["biomarkers"]])$signals
# microbiota assay counts
counts <- assays(mae[["microbiota"]])$counts

# microbiota rowData
taxtab <- rowData(mae[["microbiota"]])